refactor(tabs): make tab-panel a separate directory
