+++ src/cascader/cascader.json
@@ -4,7 +4,7 @@
     "t-icon": "../icon/icon",
     "t-popup": "../popup/popup",
     "t-tabs": "../tabs/tabs",
-    "t-tab-panel": "../tabs/tab-panel",
+    "t-tab-panel": "../tab-panel/tab-panel",
     "t-radio-group": "../radio-group/radio-group"
   }
 }
