+++ src/tab-panel/props.ts
