+++ src/tab-panel/tab-panel.json
