+++ src/tabs/README.en-US.md
@@ -6,7 +6,7 @@
 name | type | default | description | required
 -- | -- | -- | -- | --
 animation | Object | - | Typescript：`TabAnimation` `type TabAnimation = { duration: number } & Record<string, any>`。[see more ts definition](https://github.com/Tencent/tdesign-miniprogram/tree/develop/src/tabs/type.ts) | N
-custom-style `v0.25.0` | String | - | \- | N
+custom-style | String | - | `0.25.0` | N
 external-classes | Array | - | `['t-class', 't-class-item', 't-class-active', 't-class-track']` | N
 placement | String | top | options：left/top | N
 show-bottom-line | Boolean | true | \- | N
@@ -14,7 +14,7 @@ space-evenly | Boolean | true | \- | N
 sticky | Boolean | false | \- | N
 sticky-props | Object | - | Typescript：`StickyProps`，[Sticky API Documents](./sticky?tab=api)。[see more ts definition](https://github.com/Tencent/tdesign-miniprogram/tree/develop/src/tabs/type.ts) | N
 swipeable | Boolean | true | \- | N
-theme | String | 'line' | options：line/tag/card | N
+theme | String | line | options：line/tag/card | N
 value | String / Number | - | Typescript：`TabValue` `type TabValue = string \| number`。[see more ts definition](https://github.com/Tencent/tdesign-miniprogram/tree/develop/src/tabs/type.ts) | N
 default-value | String / Number | undefined | uncontrolled property。Typescript：`TabValue` `type TabValue = string \| number`。[see more ts definition](https://github.com/Tencent/tdesign-miniprogram/tree/develop/src/tabs/type.ts) | N
 
@@ -31,8 +31,9 @@ scroll | `(scrollTop: number, isFixed: boolean)` | \-
 name | type | default | description | required
 -- | -- | -- | -- | --
 badge-props | Object | null | \- | N
+custom-style | String | - | `0.25.0` | N
 destroy-on-hide | Boolean | true | \- | N
 disabled | Boolean | false | \- | N
 label | String | - | \- | N
 panel | String / Slot | - | \- | N
-value | String / Number | - | Typescript：`TabValue` | N
+value | String / Number | - | Typescript：`TabValue`，[Tabs API Documents](./tabs?tab=api)。[see more ts definition](https://github.com/Tencent/tdesign-miniprogram/tree/develop/src/tab-panel/type.ts) | N
