+++ src/tabs/README.md
@@ -13,7 +13,7 @@ isComponent: true
 ```json
 "usingComponents": {
   "t-tabs": "tdesign-miniprogram/tabs/tabs",
-  "t-tab-panel": "tdesign-miniprogram/tabs/tab-panel"
+  "t-tab-panel": "tdesign-miniprogram/tab-panel/tab-panel"
 }
 ```
 
@@ -109,7 +109,7 @@ Page({
 名称 | 类型 | 默认值 | 说明 | 必传
 -- | -- | -- | -- | --
 animation | Object | - | 动画效果设置。其中 duration 表示动画时长。TS 类型：`TabAnimation` `type TabAnimation = { duration: number } & Record<string, any>`。[详细类型定义](https://github.com/Tencent/tdesign-miniprogram/tree/develop/src/tabs/type.ts) | N
-custom-style `v0.25.0` | String | - | 自定义组件样式 | N
+custom-style | String | - | `0.25.0`。自定义组件样式 | N
 external-classes | Array | - | 组件类名，分别用于设置 组件外层元素、选项卡单项、选项卡激活态、滚动条样式类名 等类名。`['t-class', 't-class-item', 't-class-active', 't-class-track']` | N
 placement | String | top | 选项卡位置。可选项：left/top | N
 show-bottom-line | Boolean | true | 是否展示底部激活线条 | N
@@ -117,7 +117,7 @@ space-evenly | Boolean | true | 选项卡头部空间是否均分 | N
 sticky | Boolean | false | 是否开启粘性布局 | N
 sticky-props | Object | - | 透传至 Sticky 组件。TS 类型：`StickyProps`，[Sticky API Documents](./sticky?tab=api)。[详细类型定义](https://github.com/Tencent/tdesign-miniprogram/tree/develop/src/tabs/type.ts) | N
 swipeable | Boolean | true | 是否可以滑动切换 | N
-theme | String | 'line' | 标签的样式。可选项：line/tag/card | N
+theme | String | line | 标签的样式。可选项：line/tag/card | N
 value | String / Number | - | 激活的选项卡值。TS 类型：`TabValue` `type TabValue = string \| number`。[详细类型定义](https://github.com/Tencent/tdesign-miniprogram/tree/develop/src/tabs/type.ts) | N
 default-value | String / Number | undefined | 激活的选项卡值。非受控属性。TS 类型：`TabValue` `type TabValue = string \| number`。[详细类型定义](https://github.com/Tencent/tdesign-miniprogram/tree/develop/src/tabs/type.ts) | N
 
@@ -133,10 +133,10 @@ scroll | `(scrollTop: number, isFixed: boolean)` | 页面滚动时触发
 
 名称 | 类型 | 默认值 | 说明 | 必传
 -- | -- | -- | -- | --
-custom-style | String | - | 自定义组件样式 | N
 badge-props | Object | null | 透传至 Badge 组件 | N
+custom-style | String | - | `0.25.0`。自定义组件样式 | N
 destroy-on-hide | Boolean | true | 【实现有误，暂不支持】选项卡内容隐藏时是否销毁 | N
 disabled | Boolean | false | 是否禁用当前选项卡 | N
 label | String | - | 选项卡名称 | N
 panel | String / Slot | - | 用于自定义选项卡面板内容 | N
-value | String / Number | - | 选项卡的值，唯一标识。TS 类型：`TabValue` | N
+value | String / Number | - | 选项卡的值，唯一标识。TS 类型：`TabValue`，[Tabs API Documents](./tabs?tab=api)。[详细类型定义](https://github.com/Tencent/tdesign-miniprogram/tree/develop/src/tab-panel/type.ts) | N
