+++ src/tabs/_example/with-badge/index.json
@@ -2,6 +2,6 @@
   "component": true,
   "usingComponents": {
     "t-tabs": "tdesign-miniprogram/tabs/tabs",
-    "t-tab-panel": "tdesign-miniprogram/tabs/tab-panel"
+    "t-tab-panel": "tdesign-miniprogram/tab-panel/tab-panel"
   }
 }
