+++ src/tabs/index.ts
@@ -1,0 +1,3 @@
+export * from './tabs';
+export * from './type';
+export * from './props';
