+++ src/tabs/type.ts
@@ -78,7 +78,7 @@ export interface TdTabsProps {
   };
   /**
    * 标签的样式
-   * @default 'line'
+   * @default line
    */
   theme?: {
     type: StringConstructor;
@@ -95,63 +95,6 @@ export interface TdTabsProps {
    * 激活的选项卡值，非受控属性
    */
   defaultValue?: {
-    type: null;
-    value?: TabValue;
-  };
-}
-
-export interface TdTabPanelProps {
-  /**
-   * 透传至 Badge 组件
-   * @default null
-   */
-  badgeProps?: {
-    type: ObjectConstructor;
-    value?: object;
-  };
-  /**
-   * 自定义组件样式
-   * @default ''
-   */
-  customStyle?: {
-    type: StringConstructor;
-    value?: string;
-  };
-  /**
-   * 选项卡内容隐藏时是否销毁
-   * @default true
-   */
-  destroyOnHide?: {
-    type: BooleanConstructor;
-    value?: boolean;
-  };
-  /**
-   * 是否禁用当前选项卡
-   * @default false
-   */
-  disabled?: {
-    type: BooleanConstructor;
-    value?: boolean;
-  };
-  /**
-   * 选项卡名称
-   * @default ''
-   */
-  label?: {
-    type: StringConstructor;
-    value?: string;
-  };
-  /**
-   * 用于自定义选项卡面板内容
-   */
-  panel?: {
-    type: StringConstructor;
-    value?: string;
-  };
-  /**
-   * 选项卡的值，唯一标识
-   */
-  value?: {
     type: null;
     value?: TabValue;
   };
